backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0

python-dotenv==1.1.0

psycopg2-binary==2.9.11
pgvector==0.4.1

ollama==0.5.3

fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
